fix: default port of distro is 8088
