+++ clients/spring-boot-starter-camunda-sdk/src/main/resources/modes/self-managed.yaml
@@ -4,6 +4,5 @@ camunda:
     auth:
       token-url: http://localhost:18080/auth/realms/camunda-platform/protocol/openid-connect/token
       audience: zeebe-api
-    enabled: true
     grpc-address: http://localhost:26500
     rest-address: http://localhost:8088
